Move are_all_users_on_domain checks to main data store.
